Extract logic in TargetContentHasher
