+++ Sources/TuistCache/ContentHashing/GraphContentHasher.swift
@@ -9,32 +9,14 @@ public protocol GraphContentHashing {
 }
 
 public final class GraphContentHasher: GraphContentHashing {
-    private let contentHasher: ContentHashing
-    private let coreDataModelsContentHasher: CoreDataModelsContentHashing
-    private let sourceFilesContentHasher: SourceFilesContentHashing
-    private let targetActionsContentHasher: TargetActionsContentHashing
+    private let targetContentHasher: TargetContentHashing
 
     // MARK: - Init
 
-    public convenience init(contentHasher: ContentHashing = CacheContentHasher()) {
-        self.init(
-            contentHasher: contentHasher,
-            sourceFilesContentHasher: SourceFilesContentHasher(contentHasher: contentHasher),
-            targetActionsContentHasher: TargetActionsContentHasher(contentHasher: contentHasher),
-            coreDataModelsContentHasher: CoreDataModelsContentHasher(contentHasher: contentHasher)
-        )
-    }
-
     public init(
-        contentHasher: ContentHashing,
-        sourceFilesContentHasher: SourceFilesContentHashing,
-        targetActionsContentHasher: TargetActionsContentHashing,
-        coreDataModelsContentHasher: CoreDataModelsContentHashing
+        targetContentHasher: TargetContentHashing = TargetContentHasher()
     ) {
-        self.contentHasher = contentHasher
-        self.sourceFilesContentHasher = sourceFilesContentHasher
-        self.coreDataModelsContentHasher = coreDataModelsContentHasher
-        self.targetActionsContentHasher = targetActionsContentHasher
+        self.targetContentHasher = targetContentHasher
     }
 
     // MARK: - GraphContentHashing
@@ -47,7 +29,9 @@ public final class GraphContentHasher: GraphContentHashing {
                 return nil
             }
         }
-        let hashes = try hashableTargets.map { try hash(targetNode: $0) }
+        let hashes = try hashableTargets.map {
+            try targetContentHasher.contentHash(for: $0)
+        }
         return Dictionary(uniqueKeysWithValues: zip(hashableTargets, hashes))
     }
 
@@ -61,37 +45,5 @@ public final class GraphContentHasher: GraphContentHashing {
         let cacheable = isFramework && noXCTestDependency && allTargetDependenciesAreHasheable
         visited[target] = cacheable
         return cacheable
-    }
-
-    private func hash(targetNode: TargetNode) throws -> String {
-        let target = targetNode.target
-        let sourcesHash = try sourceFilesContentHasher.hash(sources: target.sources)
-        let resourcesHash = try hash(resources: target.resources)
-        let coreDataModelHash = try coreDataModelsContentHasher.hash(coreDataModels: target.coreDataModels)
-        let targetActionsHash = try targetActionsContentHasher.hash(targetActions: target.actions)
-        let stringsToHash = [sourcesHash,
-                             target.name,
-                             target.platform.rawValue,
-                             target.product.rawValue,
-                             target.bundleId,
-                             target.productName,
-                             resourcesHash,
-                             coreDataModelHash,
-                             targetActionsHash]
-        // TODO: hash headers, platforms, version, entitlements, info.plist, target.environment, target.filesGroup, targetNode.settings, targetNode.project, targetNode.dependencies ,targetNode.target.dependencies
-
-        return try contentHasher.hash(stringsToHash)
-    }
-
-    private func hash(headers: Headers) throws -> String {
-        let hashes = try (headers.private + headers.project + headers.project).map { path in
-            try contentHasher.hash(fileAtPath: path)
-        }
-        return try contentHasher.hash(hashes)
-    }
-
-    private func hash(resources: [FileElement]) throws -> String {
-        let hashes = try resources.map { try contentHasher.hash(fileAtPath: $0.path) }
-        return try contentHasher.hash(hashes)
     }
 }
